Add missing files

+++ src/Systems/DamageSystem.h
@@ -21,8 +21,7 @@ class DamageSystem: public System {
         void OnCollision(CollisionEvent& event) {
             Entity a = event.a;
             Entity b = event.b;
-            spdlog::info("Collision event emitted: {} and {}", std::to_string(a.GetId()) , std::to_string(b.GetId()));
-
+        
             if (a.BelongsToGroup("projectiles") && b.HasTag("player")) {
                 OnProjectileHitsPlayer(a, b); // "a" is the projectile, "b" is the player
             }
@@ -34,16 +33,18 @@ class DamageSystem: public System {
             if (a.BelongsToGroup("projectiles") && b.BelongsToGroup("enemies")) {
                 OnProjectileHitsEnemy(a, b); // "a" is the projectile, "b" is the enemy
             }
-
+            
             if (b.BelongsToGroup("projectiles") && a.BelongsToGroup("enemies")) {
-                OnProjectileHitsEnemy(b, a); // "a" is the projectile, "b" is the enemy
+                OnProjectileHitsEnemy(b, a); // "b" is the projectile, "a" is the enemy
             }
         }
 
         void OnProjectileHitsPlayer(Entity projectile, Entity player) {
             const auto projectileComponent = projectile.GetComponent<ProjectileComponent>();
 
-            if (!projectileComponent.isFriendly) {
+            if (!projectileComponent.isFriendly && player.HasComponent<HealthComponent>()) {
+                Logger::Log("Projectile id: " + std::to_string(projectile.GetId()) + " collided with player id: " + std::to_string(player.GetId()));
+
                 // Reduce the health of the player by the projectile hitPercentDamage
                 auto& health = player.GetComponent<HealthComponent>();
 
@@ -63,19 +64,21 @@ class DamageSystem: public System {
         void OnProjectileHitsEnemy(Entity projectile, Entity enemy) {
             const auto projectileComponent = projectile.GetComponent<ProjectileComponent>();
 
-            // Only damage the enemy if projectile is friendly
-            if (projectileComponent.isFriendly) {
+            if (projectileComponent.isFriendly && enemy.HasComponent<HealthComponent>()) {
+                Logger::Log("Projectile id: " + std::to_string(projectile.GetId()) + " collided with enemy id: " + std::to_string(enemy.GetId()));
+                
+                // Reduce the health of the enemy by the projectile hitPercentDamage
                 auto& health = enemy.GetComponent<HealthComponent>();
 
-                // Subtract from enemy health
+                // Subtract the health of the enemy
                 health.healthPercentage -= projectileComponent.hitPercentDamage;
 
-                // Kills the enemy if health reaches zero
+                // Kills the enemy when health reaches zero
                 if (health.healthPercentage <= 0) {
                     enemy.Kill();
                 }
 
-                // Destroy projectile
+                // Kill the projectile
                 projectile.Kill();
             }
         }
